Make log runner code reusable
