+++ aioesphomeapi/log_reader.py
@@ -7,16 +7,11 @@ import logging
 import sys
 from datetime import datetime
 
-import zeroconf
 
-from aioesphomeapi.api_pb2 import SubscribeLogsResponse  # type: ignore
-from aioesphomeapi.client import APIClient
-from aioesphomeapi.core import APIConnectionError
-from aioesphomeapi.model import LogLevel
-from aioesphomeapi.reconnect_logic import ReconnectLogic
+from .api_pb2 import SubscribeLogsResponse  # type: ignore
+from .client import APIClient
 
-_LOGGER = logging.getLogger(__name__)
-
+from .log_runner import async_run_logs
 
 async def main(argv: list[str]) -> None:
     parser = argparse.ArgumentParser("aioesphomeapi-logs")
@@ -42,41 +37,17 @@ async def main(argv: list[str]) -> None:
     )
 
     def on_log(msg: SubscribeLogsResponse) -> None:
-        time_ = datetime.now().time().strftime("[%H:%M:%S]")
-        text = msg.message
-        print(time_ + text.decode("utf8", "backslashreplace"))
+        time_ = datetime.now()
+        message: bytes = msg.message
+        text = message.decode("utf8", "backslashreplace")
+        print(f"[{time_.hour:02}:{time_.minute:02}:{time_.second:02}]{text}")
 
-    has_connects = False
-
-    async def on_connect() -> None:
-        nonlocal has_connects
-        try:
-            await cli.subscribe_logs(
-                on_log,
-                log_level=LogLevel.LOG_LEVEL_VERY_VERBOSE,
-                dump_config=not has_connects,
-            )
-            has_connects = True
-        except APIConnectionError:
-            await cli.disconnect()
-
-    async def on_disconnect(  # pylint: disable=unused-argument
-        expected_disconnect: bool,
-    ) -> None:
-        _LOGGER.warning("Disconnected from API")
-
-    logic = ReconnectLogic(
-        client=cli,
-        on_connect=on_connect,
-        on_disconnect=on_disconnect,
-        zeroconf_instance=zeroconf.Zeroconf(),
-    )
-    await logic.start()
+    stop = await async_run_logs(cli, on_log)
     try:
         while True:
             await asyncio.sleep(60)
     except KeyboardInterrupt:
-        await logic.stop()
+        await stop()
 
 
 if __name__ == "__main__":
